requirements.txt
Django>=5.2.3,<6.0.0
requests>=2.32.3,<3.0.0
python-dotenv>=1.0.1,<2.0.0
psycopg2-binary>=2.9.10,<3.0.0
Pillow>=11.2.1,<12.0.0
ipython>=9.3.0,<10.0.0
redis>=6.2.0,<7.0.0
djangorestframework>=3.16.1,<4.0.0
django-filter>=25.1,<26.0
djangorestframework-simplejwt>=5.5.1,<6.0.0
coverage>=7.10.6,<8.0.0
drf-spectacular>=0.28.0,<0.29.0
django-cors-headers>=4.7.0,<5.0.0
stripe>=12.5.1,<13.0.0
celery>=5.5.3,<6.0.0
eventlet>=0.40.3,<0.41.0
django-celery-beat>=2.8.1,<3.0.0
drf-yasg>=1.21.11,<2.0.0
flake8==7.3.0
whitenoise==6.6.0
gunicorn
django-redis